[kernel] Replace uint32_t with more suitable types

+++ kernel/include/libkern/mem.h
@@ -3,13 +3,13 @@
 
 #include <libkern/types.h>
 
-void* memset(void* dest, uint8_t fll, uint32_t nbytes);
-void* memcpy(void* dest, const void* src, uint32_t nbytes);
-void* memccpy(void* dest, const void* src, uint8_t stop, uint32_t nbytes);
-void* memmove(void* dest, const void* src, uint32_t nbytes);
-int memcmp(const void* src1, const void* src2, uint32_t nbytes);
+void* memset(void* dest, uint8_t fll, size_t nbytes);
+void* memcpy(void* dest, const void* src, size_t nbytes);
+void* memccpy(void* dest, const void* src, uint8_t stop, size_t nbytes);
+void* memmove(void* dest, const void* src, size_t nbytes);
+int memcmp(const void* src1, const void* src2, size_t nbytes);
 
-char* kmem_bring_to_kernel(const char* data, uint32_t size);
-char** kmem_bring_to_kernel_ptrarr(const char** data, uint32_t size);
+char* kmem_bring_to_kernel(const char* data, size_t size);
+char** kmem_bring_to_kernel_ptrarr(const char** data, size_t size);
 
 #endif // _KERNEL_LIBKERN_MEM_H
